Switching to ints instead of longs.  Can't run for graphs > 12 or so without blowing the heap because of my way of managing queues.  Need to go read some papers to figure out how to backtrack without 2^n space complexity

+++ src/main/java/com/github/steveash/maxclique/Graph.java
@@ -48,12 +48,12 @@ public class Graph<T> {
         return neighbors;
     }
 
-    public long neighborsExcludingAsMask(int vertex, long toExclude) {
-        long mask = 0;
+    public int neighborsExcludingAsMask(int vertex, int toExclude) {
+        int mask = 0;
         for (int i = 0; i < verticies.size(); i++) {
             if (vertex == i) continue;              // skip myself
 
-            long candidate = (1 << i);
+            int candidate = (1 << i);
             if ((candidate & ~toExclude) == 0) continue;
 
             if (weight(vertex, i) > 0) {
